requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
amqp==5.3.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
datasets==4.0.0
dill==0.3.8
filelock==3.18.0
Flask==2.3.3
frozenlist==1.7.0
fsspec==2025.3.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
numpy==1.26.4
packaging==25.0
pandas==2.3.1
prompt_toolkit==3.0.51
propcache==0.3.2
pyarrow==20.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
six==1.17.0
spotipy==2.23.0
sympy==1.14.0
tokenizers==0.15.2
torch==2.2.2
torchaudio==2.2.2
torchcodec==0.0.0.dev0
tqdm==4.67.1
transformers==4.35.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
